Skip configAndClose When Camera Unavailable
If the CameraDevice is currently in use by another
application, we should not attempt to perform configAndClose.
This could inadvertently disconnect the camera for the other app.

While skipping configAndClose might leave a configured Surface
connected to the camera, the other app would ideally disconnect
any stale Surfaces from the previous session when it takes control
of the camera device.

Therefore, it should be safe to ignore configAndClose in this scenario.

Test: CameraDisconnectTest
Change-Id: Ifcaa348df99c089d5714e27292a0f00cd814e37b