Support Android 14's non-linear font scaling for Material Text
The platform converter API is not public yet. As a result the code is copied from AOSP so that it can be used directly. In future we will switch to the platform API when it becomes available.

Relnote: Non-scalable texts in Material library now work correctly with Android 14's non-linear font scaling.

Bug: 292092112
Change-Id: I6601e2af6f197ac7da8407a43349bad799fecbc5