webkit: add missing @UiThread and thread checks.
All of these methods have a thread check in the framework implementation
and thus must be called on the UI thread. Unfortunately many of them did
*not* check the thread on older OS versions that don't support the API
in the framework, but it seems extremely unlikely that apps are relying
on this, so add the missing thread checks.

Bug: 304706627
Test: n/a
Relnote: Methods are now documented with @UiThread where appropriate.
Change-Id: I44541931ef0aed0a9755de353b098aa96d0c2604