Simplify "RemoteListDelegate" naming, and add `@RestrictTo(LIBRARY)`
For naming, we drop "Remote" or "Delegate", because the two are
redundant. It is safe to update the name because all relevant APIs are
marked as @RestrictTo(LIBRARY), or @ExperimentalCarApi, both of which
indicate an unstable API.

For @RestrictTo(LIBRARY), please notice that ListDelegate is only
intended for consumption by the A4C host (which is developed and managed
by Google). While there is no harm in exposing it, overexposure could be
confusing for developers.

Bug: 343220772
Test: ./gradlew :car:app:app:test
Relnote: N/A
Change-Id: I311816e60feed676cdb835767673788d7473a992