Unrestrict Biometric Error and Result Types
This removes the restrictions on the AuthenticationError and the
AuthenticationResult types, so that these can be added to the public
API. These are expected to be used by providers in response objects,
packaged through CredentialManager.

The AuthenticationError:

Error returned from the Biometric Prompt flow that is executed
by [androidx.credentials.CredentialManager] after the user
makes a selection on the Credential Manager account selector.

The AuthenticationResult:

Successful result returned from the Biometric Prompt authentication
flow handled by [androidx.credentials.CredentialManager].

The BiometricPromptResult wrap these errors and results.

Then, the ProviderCreateCredentialRequest and the ProviderGetCredentialRequest
allow retrieval of this Result type through known Jetpack APIs.

Bug: 339650358
Test: Test Cases in JUnit cover this change + Build
Relnote: "Adds the ability for providers to retrieve the error/result
from the optional Biometric flow handled by CredentialManager"

Change-Id: I4fd5ca5a6b5c0ff75ec72ef0d066162983c96944