Disable prefetch on Robolectric
Value we are getting from Choreographer.doFrame() callback:on Robolectric in GraphicsMode.Mode.NATIVE is incorrect. The values between it and System.nanoTime() should be consistent for the prefetch to work as expected.
Example of the values we get on Android Emulator:
23_673_695_346_559 Choreographer.doFrame()
23_673_701_094_286 System.nanoTime()
Example of the values we get on Robolectric
480_430_000_000 Choreographer.doFrame()
24_424_536_043_875 System.nanoTime()
With those numbers we just constantly schedule extra frame with a hope to be able to have enough available time in the next frame.
Disabling prefetch should be safe enough for tests, as it is mostly a performance related optimization, and no-one is doing performance testing on Robolectric anyway.

Bug: 339343119
Fixes: 341880461
Test: added PrefetchNotHangingMainThreadTest
Change-Id: Ib921b5f7032015866183a8136be615c6d9a8d0b8