Handle empty CollectionNavType
For collection types, in the past if no value was passed in, the arg would be ignored and omitted from argument bundle. Now, a missing collection type value will default to the NavArgument's defaultValue, or an emptyList if there isn't a defaultValue.

What is considered "no value" is different between list of strings and list of non-strings. For clarity, this is how various routes are handled:

Given a route pattern "www.test.com?list={myList}"

List<String>
1. "www.test.com?list=" considered as listOf("")
2. "www.test.com?list={myList}" considered as listOf("{myList}")
3. "www.test.com?" considered as no value
4. "www.test.com" considered as no value

non-String lists
1. "www.test.com?list=" considered as no value (since it is invalid)
2. "www.test.com?list={myList}" considered as no value
3. "www.test.com?" considered as no value
4. "www.test.com" considered as no value

Test: ./gradlew navigation:navigation-common:cC
Test: ./gradlew navigation:navigation-runtime:cC
Bug: 341723133
Relnote: "Fix how navigation handles missing arg values for CollectionNavTypes - instead of ignoring missing values, navigation will now default to its defaultValue, or an empty collection if there isn't a defaultValue. This fallback value will be added the NavBackStackEntry.argument bundle."

Change-Id: Id630fb148782fbb34d3c236f5d8bb363fba7ef91