[KSP2] Update test for Java error types
KSP2 keeps the nullability for error types. This should be a good change as more information is kept. For example the nullability of `SomeErrorType?` is nullable. When the origin is Java the nullability becomes Platform and since XProcessing now treats most Platform types as nullable we're updating our tests to reflect that.

Test: XTypeTest
Bug: 314151707
Change-Id: Ib05f467be78421859d9e3f45724d84f4af1fc1b4