Clean up when seekable transition is disposed
This CL ensures the clean-up on dispose also covers Seekable
transition case. As a part of the clean-up, the scope from
the (disposed) transition is removed from the SnapshotStateObserver,
which outlives the transition. This change effectively avoids
leaking transition state, as well as any user state that is used
as target states, such as context.

Bug: 336167334
Test: new test added

Change-Id: I89953ff6338e468ef88dfc06e0cb02e88ef8405a