Update window-demos to target API level 35.
This CL updates the window-demos projects to target API level 35.
This includes updating the build.gradle files to use compileSdk 35
and targetSdkVersion 35, as well as updating the styles.xml file to
opt-out edge-to-edge display mode.

This CL also includes a few other minor changes, such as updating the
Desugar JDK Libs to 2.0.3 to get rid of "member always wins" problem
(go/sequenced-collections-android-v-kotlin) and deprecating
the onLowMemory() callback in WindowStateCallbackActivity.

Fix: 343301487
Test: Manual launch each demo app to check window insets
Test: Run presubmit v2/androidx-core/emulator_mediumphone_api34
Test: Run presubmit v2/androidx-core/emulator_pixel2_api33
Change-Id: I8d394029ea586736b79a4f1e6e76cf411c5fb4e3