Abstract loadBannerAd method for SDK
The loadBannerAd method takes AdType and SdkType
as parameters and loads the bannerAd of specific
type eg. Mediated/Non-Mediated, WebView etc.
and also takes the parameter to wait inside
the onDraw method of the SdkView. This method
can be used to load different types of Ads
from different kinds of SDKs.

Bug: 346516922
Test: UiLib Manual Test
Change-Id: I95d7a87514d89eb5dec46474b93b53b24e5f9d46