Prevent unintended usage of TypeMirror.toString
This method, used in code processors, includes type-use annotations in the returned string. That hasn't mattered up until now since we don't use type-use annotations in AndroidX, but in preparing to enable JSpecify annotations I found a couple bugs in versioned parcelable (see aosp/3092119 for one example) where type-use annotations caused errors.

This adds a lint issue to detect usages of TypeMirror.toString, with an autofix to use TypeName.toString instead, which does not contain annotations. It also replaces existing usages of TypeMirror.toString. The existing usages all either compare the type string to unannotated string (e.g. "int" or a class name) or are used for casts in generated code, and it doesn't make sense to use the annotated string in those cases.

Bug: 339204110
Test: TypeMirrorToStringTest, this also fixes a versionedparcelable error found while testing jspecify
Change-Id: Ia3e4ea30d2d8e243a732c7eba93aa0b307748a29