Add test to prevent HorizontalPager regressions on Wear
Successfully identified the introduction and resolution of the issue
by combining this test with a binary search through past AndroidX
builds. For each build, the local repo manifest was synced with the
corresponding manifest for the previous build to determine if the
issue was present.

Behaviour broke at aosp/3010134 (test included in this change fails)
and was fixed at aosp/3094408 (test passes). So this test should
prevent further regressions.

Bug: 345187438
Test: Included
Change-Id: I6b9c9971c9e6c8add1a30c6177350d11be991ce4