Add isIndexable field to all Item types
isIndexable will be used by the AlphaJump feature, to exclude
non-alphabetic items. The first use-case will be the "Search by Voice"
button in the built-in Dialer app, which does not benefit from
alphabetical indexing.

Note that all new APIs are marked as experimental

Note that this change is compatible with MenuTemplate and
SectionedItemTemplate, which offer host-side alphajump (experimental
feature)

Note that isIndexable was not added to the base Item class. We should
consider moving common functionality into a base class in the future
(isIndexable, and other fields like title, icon, etc.).

Bug: 343726020
Test: ./gradlew :car:app:app:test
Relnote: N/A
Change-Id: Ia46e48cdfda1670b9c62a9325ac92f6fdc4f3640