Set a busy timeout during connection configuration in Room
Setting a busy timeout helps with slow IO or auto-checkpoint locks in a multiple connection database. The amount of time set is somewhat arbitrary. Its based off the 2500 ms that Android bindings already use and is meant to handle small table locks but not too big to make operations hang and apps to feel unresponsive. If a busy timeout is already configured then Room will not override it, sometimes a developer might want to set their own timeout if their database is used by multiple processes frequently and concurrently.

Bug: 304295573
Test: BuilderTest
Change-Id: I4447861b285700e6359faf6353fd8564a629263a