Fix `onForgotton` order with `nonSkippingGroupOptimization`
When `nonSkippingGroupOptimization` is enabled, the slot table order
is not good enough to order the `onForgotton` calls because remember
now adds slots to the group it is called in instead of adding them
to separate group. The runtime was modified to re-order the calls to
`onForgotton` in these cases.

However, if the remember's happend at the end of a group, the compiler
may not create a group around the calls to remember and the runtime
needs to detect when remember was not called without the benefit of
a group. This handles cases where an application is built with
modules are compile with the feature enabled and some  without this
feature enabled.

This fixes this issue by ensuring when slots are "trimmed" from
a group the `onForgotton` reordering infomration is generated for
slots added after children cals. Also, detecting when to trim
was moved ahead of removing children to ensure the `onForgotton`
order is correct.

Test: ./gradlew :compose:r:r:tDUT
Change-Id: I62596fe7ad673adf14557ef891783f96986517e6