Fix SharedTransitionObserver leaking scopes
This change adds a diposed flag to the SharedTransitionScope
to ensure after the scope is diposed, no observations can
be made on the shared transition scope or on shared elements
created in that scope. This prevents scopes from being
added to the observer after the they are disposed, therefore
preventing leaks.

Fixes: 347035242
Test: new test added
Change-Id: I8956a5c3d1ba25d0bb4600bd6f3975b2ef301f3d