Avoid allocating an new array unnecessarily in commit
Use the singleton emptyList() instead when the segment is empty. (ParamQueue.release returns a MutableCollection, so even when it's empty, that returns the ArrayList it was building and constructs a new one.) This avoids some GC e.g. when a draw loop is repeatedly calling commit. Also changes an existing usage to use kotlin.collections.emptyList() instead of java.util.Collections.emptyList(). Shouldn't make a big difference, but it's easier to structure the tests if it's using the same singleton, and the former seems more idiomatic for Kotlin code.

A new test asserts that both the implicit draw on surface creation and an empty commit use that singleton.

Test: New unit-test
Change-Id: Ife1532b66d9ff96839df8373fc0b942d2878f5a7