Improve error message on field-declared serializers
Safe Args implementation currently does not support serializers declared on a class field. This applies regardless of whether the field Type is a custom type or third-party type.

Improve error messaging to clarify that this is an unsupported feature as of 2.8.0.

This issue is expected to be addressed in navigation 2.9.

Test: ./gradlew navigation:navigation-common:test
Bug: 341319151
Relnote: "Improve error message for customer serializers declared directly on class fields via @Serializable(with =...) to clarify that this is currently an unsupported feature."
Change-Id: I052b0e6a34c7e38520723e7c00a920dc7abe4eb7