Move entries to proper state when finishing predictive back
For Navigation Compose, when doing a predictive back gesture, if the
gesture is cancelled, the NavBackStackEntry never gets its Lifecycle
RESUMED again.

Ideally, the `animateTo()` API on `SeekableTransitionState` would handle
the cancel case for us, but it restarts the animation instead of
continuing from the most recent state of the animation. So navigation
needs to handle this case explicitly and we can remove the fix once
`animateTo()` does handle it.

RelNote: "Fixed an issue in Navigation Compose where after cancelling
the Predictive Back Gesture, the NavBackStackEntry that the user returns
to never moves back to RESUMED."
Test: added NavHostPredictiveBackTest
Bug: 346608857

Change-Id: I97a0c5fe7fe661a2fadfbb01346676482d426028