Fix the extensions issue that camera stops working after suspend the app and resume on Xiaomi 13
After migrating to the new extensions capture pipeline,
we no longer create temporary YUV surface
for CaptureProcessorImpl#onOutputSurface when opening the camera.
Instead, we pass the capture surface from the ImageCapture to the
onOutputSurface directly.

On Xiaomi 13 or 13T devices, it doesn't close the previous ImageWriter
that connects to the output surface when camera is closed.
Therefore when suspend the app and resume, same ImageCapture surface
is passed to the onOutputSurface and cause it failed to connect the
surface to the another ImageWriter.

To ensure the smoothly migration to the new capture pipeline,
we do the workaround in CameraX by having a intermedia surface in between.

Bug: 347142571
Test: LifecycleStatusChangeStressTest
Change-Id: I33c36899d5289a2ab5364ce872488452581351a5