Add session management to SharedUiContainer
Adds session management logic to SharedUiContainer to maintain a
communication channel between the client and UI providers (SDKs), remote
or local. The container allows to place both client-owned and
provider-owned (via SandboxedSdkView) UI in it, which needs to be
registered with their asset IDs. For provider-owned views, corresponding
adapters can be provided. Assets can be registered before or after the
session is open.

More details on the container's behaviour can be found in the
SharedUiContainer KDoc.

Relnote: "Added session management and asset registration logic to SharedUiContainer. Known issues: when placed inside a PoolingContainer, SharedUiContainer closes the session on window detachment."
Test: build, SharedSessionIntegrationTests, SharedUiContainerTest && UI Lib Manual test app
Bug: 263460954

Change-Id: Ic60b02e354b9b561732e914599269e06050405e5