Mark an AndroidView test as flaky
Devices in CI seem to be introducing extra unexpected callbacks that
we can't reliably filter or predict. This CL splits the flaky test into
two: one that must pass and asserts a lower bound, and one that will
likely flake which asserts an upper bound. The former is kept unmarked
and must always pass for our implementation to be correct. The latter
is focused more on performance and has less of an impact if it were to
fail, and is therefore allowed to flake by this commit.

Test: AndroidViewTest
Fixes: b/346413923
Change-Id: I6c977afb274071e693a41a81e51ee622dfd940c5