Fix issue with predictive back cancellation
When a Predictive back gesture that is not seekable is interrupted by
another fragment operation, fragment manager loses it state and fails to
display the proper fragment.

We should make it so that when a Predictive back gesture that is not
seekable is interrupt we still move to the correct state.

This ignores the fact that the system will continue to dispatch calls,
but that is okay, we should ignore any calls after our cancellation.

RelNote: "Fixed an issue in fragments where interrupting a predictive
back gesture would send the fragment manager into an undefined state and
even up showing the wrong fragment."
Test: modified and added tests
Bug: 338624457

Change-Id: If82e2cd540a5319faa2e40d4a28ce31026e2e19d