Remove GMS fallback from Photo Picker contracts
Rely on the public fallback action/extras
available to all OEMs and system apps to handle
the Photo Picker contract if the system API is
not available.

Relnote: "The `PickVisualMedia` and `PickMultipleVisualMedia`
Activity Result contracts that trigger the Photo Picker no
longer directly delegates to Google Play services if the system
Photo picker is not available, but uses the public
`ACTION_SYSTEM_FALLBACK_PICK_IMAGES` action and related extras
to provide a consistent Photo Picker experience to OEMs and
system apps as a fallback. This should have no effect on
the user experience for devices that have a recent version of
Google Play services."
Test: testapp still works on updated Play Store emulator
BUG: 264622388

Change-Id: I3513de5543be10405fa77cdaeb200ce94884e95c