Migrate MultiInstanceInvalidationClient to Coroutines
This change moves the invalidation client to Coroutines, aligning with the rest of the Coroutine-based Invalidation Tracker refactor for KMP. It also changes when the 'all-tables' observer is registered. The observer will now be registered when the client is created in an start(), which should always be on a background thread during InvalidationTracker.internalInit(). This reduces the likelihood of missing invalidations (b/324260327) it also keeps things simple and avoids the racing nature of onServiceConnected and onServiceDisconnected being used to register / unregister the observer (b/350927013).

This CL also greatly cleans up the MultiInstanceInvalidationTest which has been plague with flaky and inconsistent tests due to its concurrency nature. First, it removes the benchmark relates tests, they need to be properly moved to the benchmark integration project. Second it removes asserting behaviour using LiveData, since there is a lot of thread hopping it is harder to keep it deterministic yet it doesn't provide additional coverage, as long as observers are asserted to be invalidated it is expected then that observers that back LiveData would also be invalidated.

Bug: 309996304
Bug: 350927013
Bug: 324260327
Bug: 324274513
Bug: 335890993
Bug: 330519843
Bug: 186405912
Test: MultiInstanceInvalidationTest
Change-Id: I26cbc06c752da7e465dc9ec8831ef5bc5e343606