Fix AndroidFragment on the back stack
When a fragment is using ComposeView and an
AndroidFragment within it, these child fragments
restore themselves before the ComposeView
machinery is available.

This means that Fragments that are used to being
added to static containers that always exist in
the parent's layout are suddenly in the case where
they are in a 'dynamic container' - one that only
appears well after the normal lifecycle for
fragments moves beyond this phase.

Normally, if a fragment's container is not available,
we crash - you've done something wrong. However, in
the case of dynamic containers (noted by the usage of
FragmentTransaction.add that uses a ViewGroup), we
can avoid crashing and instead wait for
onContainerAvailable to be fired to attach the fragment
to the container.

By holding the fragment's lifecycle at ACTIVITY_CREATED
until the container is made available, we ensure
that well behaved fragments added to a dynamic
container have ~the same guarantees as static containers
(e.g., their fragment will have a container by the
time they are started).

Relnote: "`AndroidFragment` now properly handles cases
where the parent fragment is put on the Fragment back stack,
avoiding 'No view found for id' issues when popping back
to that fragment."
Relnote: "Fragments added via the `FragmentTransaction.add`
method that takes a `ViewGroup` now wait for
`onContainerAvailable` before progressing to `onStart()`.
This affects users of that API, such as `AndroidFragment`,
which now waits for the `AndroidFragment` to re-enter
composition before moving it through `onStart()`."
Test: newly added test
BUG: 347706985

Change-Id: I94608fed9ada006e5e601ede8adb8be81fae29a0