Support for @RawQuery in Room KMP
Support is added by creating a new API, similar to SupportSQLiteQuery that holds the SQL query string and, in the case of this new API, a function to bind arguments into a statement. The API is named RoomRawQuery and is the common type that can be used as a the single parameter of a @RawQuery annotated DAO function. The statement received in the binding lambda will throw an error if other non bind functions are invoked.

Bug: 330586815
Test: BaseQueryTest
Relnote: "Support for @RawQuery in Room KMP by adding a new API called RoomRawQuery that is similar to SupportSQLiteQuery in terms of holding into the raw SQL string and a function to bind arguments into a statement. @RawQuery annotated functions can now accept a RoomRawQuery as their single parameter."
Change-Id: Iea84484b18c0c969b4012977e737eb12c4b4f2c6