Add DynamicTypeAnimator interface
This change introduces a new interface called ProtoAnimator, which defines methods and properties to inspect and modify Animators.

The following classes
 now implement the ProtoAnimator interface:
- QuotaAwareAnimator
- QuotaAwareAnimatorWithAux

This interface allows for inspection and modification of animations, including setting new float and int values and defining a timeframe for animation. It is primarily intended for use by UI tooling in Android Studio.

Bug: 352309787
Test: QuotaAwareAnimatorWithAuxTest, QuotaAwareAnimatorTest
Change-Id: I2078c6cf9a9bca262c274a621b1cfcab7015324d