Perf/flag follow up for aosp/3156341
Adds a flag that disables this fix.

Limits the increase in number of allocations to one per `LayoutIntrinsics` by ensuring we re-use the modified `CharSequence` for every intrinsic width calculation. This required some refactoring to move some functions inside of `LayoutIntrinsics` to share this cached `CharSequence`.

Also ensures we only create a new `SpannableString` if we need to modify the original.

Test: Same as aosp/3156341
Bug: b/346918500
Change-Id: I6946735010e48893195a5db0a75901544e196d21