Add back compat support for SharedUiAdapter
To ensure SharedUiAdapter can be used on devices that don't support SDK
Runtime and where SDKs will run in the same process as client,
LocalAdapter is introduced in SharedUiAdapterFactory. It follows the
same logic as SandboxedUiAdapter's LocalAdapter and uses reflection to
allow communication between the SDK and the client as they're in
different classloaders.

Relnote: Added back compat support for SharedUiAdapter
Test: build, SharedSessionIntegrationTests
Bug: 263460954
Bug: 340476067
Change-Id: I56d7ab5ef8c1630d339b6f1f2375a8b780dc3f33