Migrate auto-close to Coroutines
This change moves AutoCloser and relevant classes to Coroutines, aligning with the rest of the Coroutine-based infra refactor for KMP. It also reduces the lock contention by using atomics and avoiding the usage of the lock object along with improving the auto-boxing of binding arguments in the AutoClosingSQLiteStatement.

This CL also adds an internal 'Watch' interface to get the uptime in millis for time tracking, enabling time movement using a fake watch for tests. By using a test watch along with a TestCoroutineDispatcher, this CL fixes the flaky tests that have plague auto-close for a while and removes the usages of Thread.sleep(), thus creating deterministic tests.

Bug: 182343970
Bug: 189775887
Bug: 190607416
Bug: 266993269
Bug: 271325600
Bug: 283959848
Bug: 309996304

Test: ./gradlew room:room-runtime:cAT
Change-Id: I8e7e243d11d2aed63d84254f94b5626b3de89394