Generalize Bundle Passing for Beta Users
Previously, we created a build-check based solution. Since then, we have
not yet dogfooded, and while we expected the next beta release to be
beta 4, Android has released a beta3.1. Given the bugfix deadline for V,
this new beta3.1 - even though it was released after the fact - did not
include our changes in the bugfix.

For beta4 onwards, and for beta3 and prior, our solution would allow
beta testers to properly utilize the single tap flow. However, the
release of 3.1 has required we workaround this issue, since most users
will utilize 3.1.

This solution aims to address both challenges, and no longer relies on
having to maintain beta versions in the codebase (thus the removal that
we planned to do when beta4 rolled around could happen now, thus
fulfilling AndroidX council recommendations early).

Bug: 353798766
Test: Across git_main, beta2.2, beta3.1 + units
Change-Id: Idf6f4bf65c0c41cfe332681de7b49209bc0aa5a1