Add MultiProcessGlanceAppWidget to support multiprocess configurations
Introduces MultiProcessGlanceAppWidget, which is the same as GlanceAppWidget
except it allows specifying a RemoteWorkerService for the session to
run in. This supports use cases of receivers that run in different
processes (so that Workers started by a particular receiver will always
run in the same process as that receiver).

However, this still maintains the requirement that all of the components
for a particular widget (receiver, worker server, GlanceComponents) run
in the same process.

Bug: 333396676
Test: tested with demo widget
Relnote: "Add MultiProcessGlanceAppWidget to support multiprocess
 configurations"

Change-Id: Idbb90ba663c1b469462077017548f3dae230906a