Use the supported preview size which is closest to the display size
In Camera2ExtensionsUtil#pickPreviewResolution(),
there is a judgement if the difference of aspect ratio between
preview size and display size is greater than 0.2f.
But if there is no supported preview size which the above result
is true, previewSizes becomes to be empty.
Then, crash occurs when calling previewSizes[0].

Therefore, use the supported preview size which is closest to
the display size if previewSizes is empty.

Bug: 354073421
Change-Id: Ibac1adb218ed30101ceef6e309dd36210881b9b8