Resolve lifecycle owner from disjoint view parents
This CL introduces the concept of a "disjoint parent," which is a view
that is associated with and effectively the parent of another view but
is not actually the view's direct parent. A View can only have a
disjoint parent if it does not have a direct parent.

Other CLs in this topic expand the concept throughout core to other
owners. With all of the changes in this set, Lifecycle, Saved State,
ViewModel, OnBackPressedListener, and ViewTreeFullyLoaded will all be
resolvable through a disjoint parent.

Test: ViewTreeLifecycleOwnerTest, ViewTreeViewModelStoreOwnerTest
Relnote: """
    ViewTree LifecycleOwners and ViewModelStoreOwners can now be
    resolved through disjoint parents of a view, such as a ViewOverlay.
    See the release notes of core or the documentation in
    `ViewTree.setViewTreeDisjointParent` for more information on
    disjoint view parents.
    """
Change-Id: I800f4616721ada959cbc8123a919dbbf199110c4