add getAvailableStartingCallEndpoints API
Previously, there was no way for clients to fetch the call audio
endpoints without starting a call.  Clients would have to use the
AudioManager to fetch the available audio devices but there is not a 1
to 1 mapping to call endpoints.

Now, clients can fetch the call audio endpoints before starting a call
and even start a call with one of those endpoints.

see go/precallaudioapi for more

Relnote: add new CallsManager API getAvailableStartingCallEndpoints
Bug: 352057844
Test: manual + e2e + unit
Change-Id: Ia8bcf70d3a8e6c9f6a96d96bdd12204aabfff3df