Flatten curve of label height animation more aggressively
If the label height delta is large, its contribution to the
height outpaces the padding value, but only temporarily. This
leads to a bulging animation as the text field is focused.

The choice in curve is arbitrary because it isn't truly part
of the animation. The only requirements are something
continuous, monotonic, and slower than linear.

Bug: b/354649086
Test: added unit test
Change-Id: I0c4bce9722a29b21a288178eec9eb5e6c624ff7f