Remove material-icons-core dependency from the material library
Copies the single icon used in ExposedDropdownMenu into the library.
This helps make sure that Jetbrains copy of compose does not need this
library.

Relnote: """material no longer adds a dependency to material-icons-core so if
your project relied on that, you will have to explicitly add that
dependency in your build.gradle[.kts] files."""

Bug: 349894318
Test: ./gradlew -p compose:material build assembleAndroidTest
Change-Id: Ie8fb35d07ce22063d5daf175927dd5a094016cae