Relax epsilon at extrema
When computing Bézier curves with an input value close to 1.0f,
numerical inaccuracies can cause us to compute results outside
of the expected domain ([0..1]). We use a special epsilon for
this that proved to still be too small for t=1.0f-1.0f.ulp
and certain Bézier curves. This change increases epsilon
slightly, and tests every pre-defined easing curve with the
lat 16 ulps before t=1.0f.

Test: EasingTest, PathEasingTest
Fixes: 354405919
Relnote: N/A
Change-Id: Ifdb50c802be2b3d71a352f1ee4f1bb58bd9d8d48