Prevent propagation of NaN values in Compose UI and Foundation.
VelocityTracker may be generating  NaN valued velocities. This can cause unexpected behavior in Foundation when we expect velocities to be valid for things like FlingBehavior. In this change I'm coercing these values to zero in both UI and foundation, but the longer term fix should be a more in depth investigation of velocity tracker.

Fixes: 352467340
Fixes: 347564725
Fixes: 339398170
Bug: 355160589
Bug: 219980303

Test: Added tests in UI and Foundation to simulate a situation where NaN velocities would be generated (same timestamp for different offset data points)
Change-Id: I188b3928106f6bf2cf793e5336cba479a74f343b