Invoke onUiContainerChanged when window visibility changes
When onWindowVisibilityChanged is triggered on a
SandboxedSdkView, also trigger viewability calculation
if necessary. This will ensure that an accurate onScreenGeometry
is sent when the activity hosting the SandboxedSdkView is paused or
resumed.

When the hosting window is not visible, an empty Rect will
be sent as the on screen geometry value.

Relnote: Invoke onUiContainerChanged when the window visibility changes
Test: SandboxedSdkViewTest, manual
Bug: 352690169
Change-Id: I541cfd2fc205a4874f7f9387dd0a85b360b46b78