Run failing tests in room-compiler-processing* with KSP1
We enabled KSP2 for all XProcessing tests in the previous commit. Here we restore the old testing infra that uses the Kotlin CLI for KSP1 and pass compiler args "-language-version=1.9" and "-api-version=1.9" to failing tests to force them to run with the old infra. We hope to remove the args and migrate these tests to KSP2 gradually in the future.

Bug: 314151707
Test: existing tests still pass

Change-Id: I60ad9465fadc33cd84f2be720c75dbcd19e6ba8c