Make ImageWithScrimPainter use Unspecified size by default.
forcedSize param has 3 distinct values

Size.Unspecified (default) - use the Component size only
null - Use the background Painter.intrinsicSize
Size(...) - use a specific size

The Size of the button should come from

1) Modifiers (developer specified)
2) Content (labels, icons)
3) UX specified minimums

And not from the background image size.

If the image must have a treatment other than scaling, the Coil rememberAsyncImagePainter allows for a contentScale.

If the Button must be the same size as the Image, then the developer should specify a size.

Bug: 354891233
Relnote: Default Chip and Button backgrounds to Size.Unspecified.
Test: Scuba screenshot tests.
Change-Id: Ic57af1323c60a68ab6df859c3ecc348163417981