ParagraphIntegrationTest: Ensure EmojiCompat instance fully loads before running all tests
In bug b/355009543, despite the EmojiCompat instance being configured an IllegalStateException is thrown in AndroidParagraphHelper.android.kt, which indicates that the instance has started loading but is not fully loaded. Therefore, we will wait for the loading state of the instance to be LOAD_STATE_SUCCEEDED before continuing with testing.

Bug: 355009543
Test: NOT A VERIFIED FIX as this test already passed locally. This bug is only happening on postsubmit configurations.
Change-Id: I59573fc77389384620544935e6006a420a6feacd