Migrate Guava shortcut binders to driver APIs
Following the same pattern as prepared result binder a new Lambda(Insert|Upsert|Delete|Update)ResultBinder was created that will generate a lambda function for the integration APIs that use the driver and take a function to act on the SQLiteStatement.

This change also consolidates duplicated code between the Insert and Upsert providers and binders, they too can be joined like the 'delete or update' ones.

Test: DaoKotlinCodeGenTest
Change-Id: Ic4f39a736a5ebe80faeca01fd5e492811dfb29bc