Pin androidx.core:core dependency to 1.13.0
Also removes some instances of BuildCompat.isAtLeastV(), because in
1.13.0, that method has not yet been updated to detect SDK_INT 35.

For the pinned dependency:
via ag/21430571
(https://android.googlesource.com/platform/frameworks/support/+/923fc8d07c43541cf558dd79daa611c07ba42bc4),
the tip-of-tree dependency was to expose
AccessibilityNodeInfoCompat.AccessibilityActionCompat.ACTION_SCROLL_IN_DIRECTION,
which appears to have released in 1.12.0.
Since then, a dependency on DifferentialMotionFlingTarget has been
added, which was released in core 1.13.0.

For receive-content-testapp/build.gradle:
It's not clear to me that there's a need for receive-content-testapp to
explicitly depend on annotation-experimental, and it was causing
dependency resolution conflicts.  With it removed, the dependency
resolves to 1.4.0 (due to transitive dependencies on the library).

If there *should* be an explicit dependency, then it should be in the
main source set, not the test source set.

Fixes: 355221446
Test: Presubmit
Change-Id: I7662ef6b5971a1437594834f44cf4655b91d670f