Do not crash on cross axis animation update
When positioning a measured list item do not crash when cross axis is defined, just don't use it. The item will be positioned normally because we use getOffset to get the correct offset from the placeable list, but since the item animator is written in a generic way we need to pass the cross axis value for the grid cases.

Test: Additional test to verify the behavior.
Fixes: 355705780
Change-Id: I5a1d08e4ecde6fa7473af343e8d7f1c48837387c