When waiting for draw, get view from ComposeRootRegistry
Instead of getting the decorView from the activity, this gets the view
through the ComposeRootRegistry, which will also work if
EmptyComposeRule is used. An EmptyComposeRule has no reference to the
activity so it won't be able to get the decorView from it either.

Also added the wait-for-draw mechanism to awaitIdle(), even though it's
not suspending.

Bug: 355076437
Bug: 302303969
Test: ran flaky test 1000 times without failing

Change-Id: Iabc7770fe0417a24dba7632e334fe383c7a9cb28