Migrate Rx query result adapters to driver APIs
Migrate Rx query adapters by using newly added helper utility functions that can let generated code act on the connection to perform the query while returning the appropriate Rx type.

For the reactive types, Flowable and Observable the coroutine-rx bridges are used to convert from Flow to Rx. While for Single and Maybe a fromCallable builder is used along with performBlocking. This is done to maintain existing behaviour where emitter thread is controlled by subscriber, where as if we use the Coroutines Rx builders in the library, they require a Coroutine context or use Dispatchers.Default causing the actual query to be performed in a background thread not controlled by the subscriber.

Test: DaoKotlinCodeGenTest
Change-Id: I4dad5b8506c366ee3e715ee9e7e3e49ce86a0dd2