Allow Dialogs to go under the status bar
Fixes: 297377093

When decorFitsSystemWindows = false and usePlatformDefaultWidth = false
a Dialog should be allowed to take up the entire screen if desired.

This CL sets "android:windowIsFloating=false" for all versions when
decorFitsSystemWindows = false. It also measures using the input size
when both decorFitsSystemWindows and usePlatformDefaultWidth are
false. This allows the Dialog to measure to the size of the screen.
It also allows the WindowInsets to properly work in the Dialog.

When a Dialog takes up the entire screen, the background for
both the status bar and the navigation bar are set to transparent
so that the Dialog can manage the background.

When targetSdk >= 35, decorFitsSystemWindows has no effect as
it is always considered "false." V devices suppot edge-to-edge
and all Dialogs must be able to handle insets.

Test: manually tested on API 21, 26, 28, 29, 30, 31, 34, 35
Test: new test for size and insets
Change-Id: I6f844b8c93a548c1e8cb64463a9b46f403bef2a4