Add snasphot observer API
Add a tooling API that allows observing when snapshots
are created, applied and disposed. This API also allows
introducing read and write observers to snasphots as they
are created.

Fixes: 293517367
Test: New tests, ./gradle :compose:r:r:tDUT
Relnote: """Add an experimental tooling API that allows observing
when a snaphot is created, applied and disposed. This API also allows
adding read and write observers to the observed snapshot. This API is
intended only for diagnostic tooling as using it in a production
application could negitively affect application performance."""

Change-Id: I22cf87fe16d981df13339196bab5d0c4c965c918