Invalidate tracker temp triggers after configuring a connection
When a connection is configured, it either might be because the new single write connection is being opened or a previous one was lost and a new one is being recreated. Since it might be that temp triggers might have been lost, when a connection is configured, the 'needSync' flag is set such that in the next transaction triggers are recreated if need be.

Test: Existing
Change-Id: I6a991a8dba685dff259e2cb73a4c2886d7b704ce