Move local publishing repo to out/repository
Prior to this change running ./gradlew core:core would publish
to rootProject.buildDir/support_repo. This requires accessing
rootProject.buildDir when configuring every androidx project and that
assumes that the root project will be configured first. This is no
longer the case when project isolation is enabled. Instead, we move
to out/repository that is a standalone directory.

Bug: 358167922
Test: ./gradlew createArchive -> output is in a new location
      ./gradlew verifyInspectorJarIsPresent -> pass
Change-Id: I309d629ec76e4ea9030b94e21e6a01ed8cf0a5f3