[Webkit] Add Url Prefetch API
This CL adds the required API surface for URL Prefetch in WebView. This
is scheduled to be released in the next version so there's no need to
add @RestrictedTo.

Relnote: " URL Prefetching in Profile: Introduced a new API in Profile to
 allow developers to trigger URL prefetching.
Prefetch Status Updates: The PrefetchCallback can be used to receive updates
 on the progress and completion status of prefetches.
Prefetch Clearing: Added an API to clear ongoing prefetches."

Test: TBD
Bug: 355430425
Change-Id: I42c5c282a7114b8b72e62b2b24f917a0f9b47123