Support onTimeout callback in SystemForegroundService
In Android 14, a foreground service of type 'short service' has time limit and in Android 15 various other foreground service types have time limits. The way the framework informs the time limit has been reached is by calling onTimeout() of the Service. If the service is not stopped in a short amount of time then an ANR is produced.

This CL adds handling for when onTimeout() is invoked on WorkManager's FGS such that the service is stopped. The actual worker is not stopped or cancelled and may or may not continue to run depending on the proc state of the app.

On a follow-up change, APIs signals might be added to let the foreground worker the service timed out.

Bug: 355634562
Test: SystemForegroundDispatcherTest
Change-Id: Ibd046586fe2851467948fdd6f70f98be4dca06b2