Add new API CustomTabsIntent.Builder#setNetwork and relevant extra flag.
This CL introduceds a new API to CustomTabs.Builder to specify the
network to be bound when launching a URL via CustomTabs(e.g. may want to
open a custom tab over a Wi-Fi network, while the default network is a
cellular connection). All URLRequests created in the future via this tab
will be bound to this target network. Per API requirement, this CL also
introduces a getter method for the bound network.

Bug: 329361366
Test: ./gradlew updateApi
Test: ./gradlew browser:browser:test
Relnote: Add new API to set/get the bound network when launching a URL over a custom tab.
Change-Id: I493e126cdc9ff645cc75d3ff5fa5ee3fb1a08f8f